pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-sk"
version = "0.1.0"
description = "基于 Python Semantic Kernel 的多智能体系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "semantic-kernel>=0.9.0",
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "azure-identity>=1.14.0",
    "dashscope>=1.14.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/multi-agent-sk"
Repository = "https://github.com/yourusername/multi-agent-sk"
Documentation = "https://multi-agent-sk.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "patterns*", "config*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=core --cov=patterns --cov-report=html --cov-report=term"
asyncio_mode = "auto"